+++ packages/runtime/src/internal/Runtime.ts
@@ -27,6 +27,7 @@ import { type AppForCheck, hasInstalledApp } from "./hasInstalledApp.js";
 import type { RuntimeContext } from "./interfaces.js";
 import { listenDevtoolsEagerly } from "./devtools.js";
 import { getV2RuntimeFromDll } from "../getV2RuntimeFromDll.js";
+import { ThemeSettings, customizeColorTheme } from "./customizeColorTheme.js";
 
 let runtime: Runtime;
 
@@ -195,7 +196,7 @@ export class Runtime {
     this.#initialized = true;
     normalizeBootstrapData(data);
     bootstrapData = data;
-    const { notification, dialog } = this.getPreseBricks();
+    const { notification, dialog } = this.#getPresetBricks();
     if (notification !== false) {
       loadNotificationService(
         notification ?? "basic.show-notification",
@@ -205,6 +206,7 @@ export class Runtime {
     if (dialog !== false) {
       loadDialogService(dialog ?? "basic.show-dialog", this.loadBricks);
     }
+    customizeColorTheme(data.settings?.misc?.theme as ThemeSettings);
   }
 
   async bootstrap(data?: BootstrapData) {
@@ -292,11 +294,10 @@ export class Runtime {
     return loadBricksImperatively(bricks, getBrickPackages());
   }
 
-  getPreseBricks() {
+  #getPresetBricks() {
     return (bootstrapData?.settings?.presetBricks ?? {}) as {
       notification?: string | false;
       dialog?: string | false;
-      uiPatch?: boolean;
     };
   }
 }