+++ packages/build-next-libs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.8](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-libs@1.0.7...@next-core/build-next-libs@1.0.8) (2023-10-18)
+
+**Note:** Version bump only for package @next-core/build-next-libs
+
+
+
+
+
 ## [1.0.7](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-libs@1.0.6...@next-core/build-next-libs@1.0.7) (2023-09-05)
 
 **Note:** Version bump only for package @next-core/build-next-libs