+++ packages/preview/CHANGELOG.md
@@ -3,6 +3,188 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.37](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.36...@next-core/preview@0.2.37) (2023-11-08)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.36](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.35...@next-core/preview@0.2.36) (2023-11-07)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.35](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.34...@next-core/preview@0.2.35) (2023-11-07)
+
+
+### Bug Fixes
+
+* refine loading bar ([da2bd08](https://github.com/easyops-cn/next-core/commit/da2bd08c53da2fbdac87476624f3793af229a558))
+
+
+
+
+
+## [0.2.34](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.33...@next-core/preview@0.2.34) (2023-11-06)
+
+
+### Bug Fixes
+
+* set ui version in docs and playground ([bed4722](https://github.com/easyops-cn/next-core/commit/bed4722221696310eebe067ea07f2d9ef0c85161))
+
+
+
+
+
+## [0.2.33](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.32...@next-core/preview@0.2.33) (2023-11-03)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.32](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.31...@next-core/preview@0.2.32) (2023-11-03)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.31](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.30...@next-core/preview@0.2.31) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.30](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.29...@next-core/preview@0.2.30) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.29](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.28...@next-core/preview@0.2.29) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.28](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.27...@next-core/preview@0.2.28) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.27](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.26...@next-core/preview@0.2.27) (2023-11-01)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.26](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.25...@next-core/preview@0.2.26) (2023-11-01)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.25](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.24...@next-core/preview@0.2.25) (2023-11-01)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.24](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.23...@next-core/preview@0.2.24) (2023-10-30)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.23](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.22...@next-core/preview@0.2.23) (2023-10-26)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.22](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.21...@next-core/preview@0.2.22) (2023-10-20)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.21](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.20...@next-core/preview@0.2.21) (2023-10-20)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.20](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.19...@next-core/preview@0.2.20) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.19](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.18...@next-core/preview@0.2.19) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.18](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.17...@next-core/preview@0.2.18) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.17](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.16...@next-core/preview@0.2.17) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
+## [0.2.16](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.15...@next-core/preview@0.2.16) (2023-10-18)
+
+**Note:** Version bump only for package @next-core/preview
+
+
+
+
+
 ## [0.2.15](https://github.com/easyops-cn/next-core/compare/@next-core/preview@0.2.14...@next-core/preview@0.2.15) (2023-10-12)
 
 **Note:** Version bump only for package @next-core/preview