+++ packages/react-runtime/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.29](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.28...@next-core/react-runtime@1.1.29) (2023-11-08)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.28](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.27...@next-core/react-runtime@1.1.28) (2023-11-06)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.27](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.26...@next-core/react-runtime@1.1.27) (2023-11-03)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.26](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.25...@next-core/react-runtime@1.1.26) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.25](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.24...@next-core/react-runtime@1.1.25) (2023-10-30)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.24](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.23...@next-core/react-runtime@1.1.24) (2023-10-20)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.23](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.22...@next-core/react-runtime@1.1.23) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.22](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.21...@next-core/react-runtime@1.1.22) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.21](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.20...@next-core/react-runtime@1.1.21) (2023-10-19)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.1.20](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.19...@next-core/react-runtime@1.1.20) (2023-10-18)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
 ## [1.1.19](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.1.18...@next-core/react-runtime@1.1.19) (2023-10-12)
 
 **Note:** Version bump only for package @next-core/react-runtime