+++ packages/test-next/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.11](https://github.com/easyops-cn/next-core/compare/@next-core/test-next@1.0.10...@next-core/test-next@1.0.11) (2023-10-18)
+
+**Note:** Version bump only for package @next-core/test-next
+
+
+
+
+
 ## [1.0.10](https://github.com/easyops-cn/next-core/compare/@next-core/test-next@1.0.9...@next-core/test-next@1.0.10) (2023-09-12)