+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,139 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.11](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.10...@next-core/theme@1.3.11) (2023-11-07)
+
+
+### Bug Fixes
+
+* do not ignore generated.css in npm ([79366d0](https://github.com/easyops-cn/next-core/commit/79366d07f35b38a1f3491e8113407c1fb12134db))
+
+
+
+
+
+## [1.3.10](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.9...@next-core/theme@1.3.10) (2023-11-07)
+
+
+### Bug Fixes
+
+* build generated.css ([9aac2fa](https://github.com/easyops-cn/next-core/commit/9aac2fa9fc1d78247fbd00bec7bade0e3a6ba143))
+* remove generated.css ([00cd626](https://github.com/easyops-cn/next-core/commit/00cd6260863c65acd8a68415153a4b3a1b11430a))
+* unify theme css variables ([33c010f](https://github.com/easyops-cn/next-core/commit/33c010f353e5687c48709f157eeab33369995ef4))
+
+
+
+
+
+## [1.3.9](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.8...@next-core/theme@1.3.9) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/theme
+
+
+
+
+
+## [1.3.8](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.7...@next-core/theme@1.3.8) (2023-11-02)
+
+**Note:** Version bump only for package @next-core/theme
+
+
+
+
+
+## [1.3.7](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.6...@next-core/theme@1.3.7) (2023-11-01)
+
+
+### Bug Fixes
+
+* add new css variables: --grid-card-gap ([ae18899](https://github.com/easyops-cn/next-core/commit/ae188995ebf0b718d6008ee42a4ebb0da64d3fc2))
+
+
+
+
+
+## [1.3.6](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.5...@next-core/theme@1.3.6) (2023-11-01)
+
+
+### Bug Fixes
+
+* fix --tabs-outlined-content-padding ([e4804c1](https://github.com/easyops-cn/next-core/commit/e4804c138c43b104e57269de060e8cad40ab4898))
+
+
+
+
+
+## [1.3.5](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.4...@next-core/theme@1.3.5) (2023-11-01)
+
+
+### Bug Fixes
+
+* add new css variables for tabs ([1b0a72c](https://github.com/easyops-cn/next-core/commit/1b0a72c22d5c7537c39ed17ba761c4dc47f22169))
+
+
+
+
+
+## [1.3.4](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.3...@next-core/theme@1.3.4) (2023-10-30)
+
+
+### Bug Fixes
+
+* set ui v8-2 on html instead of body ([0a029f9](https://github.com/easyops-cn/next-core/commit/0a029f939a215ea180fd36fc4f21223d4b5787ae))
+
+
+
+
+
+## [1.3.3](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.2...@next-core/theme@1.3.3) (2023-10-26)
+
+
+### Bug Fixes
+
+* **css:** update --slide-up-shadow ([89ccd51](https://github.com/easyops-cn/next-core/commit/89ccd51d7f5e9acad95118216d243a53ed6d9b5b))
+
+
+
+
+
+## [1.3.2](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.1...@next-core/theme@1.3.2) (2023-10-20)
+
+**Note:** Version bump only for package @next-core/theme
+
+
+
+
+
+## [1.3.1](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.3.0...@next-core/theme@1.3.1) (2023-10-20)
+
+**Note:** Version bump only for package @next-core/theme
+
+
+
+
+
+# [1.3.0](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.2.0...@next-core/theme@1.3.0) (2023-10-19)
+
+
+### Features
+
+* **theme:** update css variable ([5c76acd](https://github.com/easyops-cn/next-core/commit/5c76acde851ea9489b733bc79dc002f5068df85b))
+
+
+
+
+
+# [1.2.0](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.1.1...@next-core/theme@1.2.0) (2023-10-18)
+
+
+### Features
+
+* remove ui-patch and add version in theme ([8900f6b](https://github.com/easyops-cn/next-core/commit/8900f6bd9b7432de19c22d595720fbeefb2dcd63))
+
+
+
+
+
 ## [1.1.1](https://github.com/easyops-cn/next-core/compare/@next-core/theme@1.1.0...@next-core/theme@1.1.1) (2023-04-10)
 
 **Note:** Version bump only for package @next-core/theme