+++ packages/types/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/easyops-cn/next-core/compare/@next-core/types@1.7.0...@next-core/types@1.7.1) (2023-10-18)
+
+**Note:** Version bump only for package @next-core/types
+
+
+
+
+
 # [1.7.0](https://github.com/easyops-cn/next-core/compare/@next-core/types@1.6.1...@next-core/types@1.7.0) (2023-10-12)